Strip suffixes in magit-ref-ambiguous-p instead of magit-ref-fullname
Since [1: 8548d2017] this function removes a suffix such as "^" before
determining the full name and then appends it to the returned value
again, but that isn't what this function is supposed to be about.

As suggested in [1: 8548d2017] the suffix stripping could also be done
in the caller `magit-ref-ambiguous-p' itself, and that is what we are
doing going forward.

I have checked the other callers they are not affected or are unused,
making it okay to change the behavior.  `magit-ref-equal' is only used
by `magit-ref-eq', which is unused, while `magit-branch-delete' and
`magit--cherry-move' only call `magit-ref-ambiguous-p' with pure
refnames.

1: 8548d2017c034f1700ffb1193be1230db687d9af
   magit-ref-fullname: support revision suffixes