Favor upstreams that can be resolved to a branch
Most callers of `magit-get-upstream-branch' only care about the
upstream if it can be resolved to a branch, but until now this
function defaulted to return some representation of the upstream
even when that is not possible.

The optional VERIFY argument was added to let callers declare that
they are only interested in resolved upstreams, but most callers
that should have been adjusted to use that argument were not.

This commit removes the VERIFY argument and replaces it with a LAX
argument, which has the opposite effect.  By default this function
now returns the resolved upstream or nil.

Additionally this function is changed to use "git rev-parse --verify
--symbolic-full-name BRANCH@{upstream}" to determine the resolved
upstream.  Previously it accessed the variables "branch.<name>.remote"
and "branch.<name>.merge" and then tried to resolve that to a local or
remote-tracking branch itself.  The old approach resulted in incorrect
results in some edge cases.

The LAX argument is for callers such as the "push to the configured
upstream even if that doesn't exist yet" command, which cares about
configured but unsolvable upstreams.  These commands need further
work, but that will be done in other commits.  When LAX is non-nil,
then this function behaves exactly as it used to behave when VERIFY
was nil, including the bugs.

`magit-get-upstream-ref' now always returns the fully qualified and
resolved upstream or nil.