magit-get-upstream-remote: Change meaning of argument
Never return "." because that represents the local repository and the
local repository by definition is not a remote.  Remove the NON-LOCAL
argument, which previously had to be used to enforce that.

If the upstream configuration referred to a missing (but named) remote
or an url instead of a name, then this function used to return that.
Now it returns nil instead, except if the new ALLOW-UNNAMED argument
is non-nil, in which case the returned value may also be an url.

Also if an url is returned, then propertize that using the face `bold'
instead of `magit-branch-remote' as before.

This also changes the behavior of the only caller;
`magit-get-current-remote'.