magit-fetch-from-upstream: Fallback to current remote
This command recently learned to set the upstream before fetching.
That was a mistake.  This command is all about fetching from some
remote, preferably the upstream remote; it does not matter what
branch is the upstream branch of the current branch.  "The upstream"
is a (potentially remote) branch, not a remote (repository).

It is more important for this command to be able to fetch from "the
upstream" in as many cases as possible, than for it to know what the
upstream is based on the configuration of the current branch.

Therefore this command now falls back to fetching from the "current
remote" if "the upstream" is unknown or a local branch.  If there is
a single remote, then that always is the current remote, otherwise
if "origin" exists, then that is used.