magit-read-upstream-branch: Limit to existing branches
That means that "b u" (`magit-branch.<branch>.merge/remote') can no
longer be used to set an upstream that doesn't exist.  This has the
potential of being a disruptive change, but I think it is the right
thing to do.  The user doesn't gain anything from having configured
an usable upstream compared to not having configured it at all.

Only push commands, mainly "p u" (`magit-push-to-upstream'), are
capable of configuring an upstream that doesn't exist yet.  They do
so by first attempting to create it (by pushing obviously) and only
if that succeeds do they configure the target as the upstream.

The problem is that the preferred, and now only, way of creating the
upstream might not be very discoverable.  On the other hand, it is
not completely unreasonable that the "initial push" is done by, well,
pushing.  But that assumes that the user thinks of it that way.  If
someone uses "the upstream" in a way that differs from the workflow
suggested in the "The Two Remotes" info node, then things might be
more confusing.

While other parts of Magit have to be able to deal with missing or
outright invalid upstreams and despite the discoverability issue, it
is probably a good idea to prevent users from configuring unusable
upstreams.