Remove option magit-remote-set-if-missing
The commands that previously respected this variable now always let
the user set the upstream/push-remote, provided HEAD is not detached.
It is no longer possible to disable this feature.

This variable was also used to configure whether "remote.pushDefault"
or "branch.<name>.pushRemote" is set by the push-remote commands.
Add a new option `magit-prefer-push-default' for that purpose.