magit-log-maybe-update-blob-buffer: Change finding existing buffer
Previously we looked for a buffer where `magit-buffer-revision' is
non-nil, but we cannot keep doing that because that variable is going
to be set in other buffers too.  The new way of finding the buffer is
also not future-proof, but we are going to switch to a file-handler
anyway, so I don't want to spend too much time on this.