Change stickiness of infix arguments
Replace the option `magit-use-sticky-arguments' with two new options,
`magit-prefix-use-buffer-arguments' (for `magit-diff' and `magit-log')
and `magit-direct-use-buffer-arguments' (for diff and log suffixes,
when they are invoked directly, not from their prefix).

Previously the arguments that are in use in the respective buffer were
always used by default, even if that buffer isn't the current buffer,
now both variables default to only use those arguments if that is that
is visible in a window of the selected frame.  For direct commands I
am fairly sure this is less surprising.  For prefixes I was on the
fence.

Also see the newly added documentation.