magit-diff-visit-file: Always visit the blob for committed changes
Previously this behaved differently for the HEAD changes.  In that
case it visited the file in the working tree instead.  There was no
way to visit the HEAD blob instead.  This was inconsistent and made it
very annoying to blame code that was already modified in the working
tree.

Add a new option `magit-diff-visit-avoid-head-blob' for users who have
grown attached to the old inconsistent behavior, but I think it would
be better to learn to use "C-RET" (`magit-diff-visit-file-worktreee')
instead of "RET" when one wants to visit the worktree file.