magit-log-goto-same-commit: Split into two functions
The new function `magit-log-goto-commit-section' jumps to the section
whose value is the commit passed as an argument.

Remove `magit-log-goto-same-commit's optional and misguided DEFAULT
argument, which was used if the previously current section disappeared
during a refresh.  But that isn't the intended purpose of this
function.

These two actions are quite different, though the latter function is
implemented using the former.