magit-branch-or-commit-at-point: Default to hash if not a branch
If there is no branch at point, then this function used to default to
a "short name" for the branch at point, e.g. something like "master~2".

Revisions like that may be prettier than hashes, but their use often
leads to undesirable behavior.  For example when rewriting history,
one had to carefully avoid refreshing the revision buffer because
doing so would cause the buffer to show some other commit.

In other cases using a name is more desirable, but until we somehow
give users a choice, lets use the safer of the two options.