Remove variable magit-display-file-buffer-function
The is one step in modeling the `magit-diff-visit-file' and
`magit-diff-visit-worktree-file' families of commands after
the `find-file' family.

The next step will be to implement the missing `*-other-window'
and `*-other-frame' variants.

In other words, users can no longer configure the base variant to
not reuse the selected window.  Instead they can use the variant
that behaves the way they like.  That may involve changing key
bindings.

One inconsistency compared to the `find-file' family is preserved:
the base aka. same-window variants continue to display in another
window if a prefix argument is used.