magit-push-current-to-pushremote: Specify destination
In 839aaedad (magit-push-current-to-pushremote: Use literal ref as
refspec, 2019-04-29), we switched to using an explicit ref rather than
HEAD for the src part of the refpec to avoid a race condition.  The
description for the 'git push origin master' example in git-push's
manpage suggests that a destination of the same name is always chosen,
but this actually depends on the value of push.default.  As described
in the refspec section of that manpage, dropping the destination
behaves as 'git push origin' (i.e., pushing with no refspec).  So if
push.default is 'upstream', then 'git push origin REF' pushes to the
configured upstream of REF (falling back to the remote ref of the same
name).

Explicitly specify the destination to make this work with
'push.default=upstream' while still avoiding the race condition.
Specifying the destination is a conservative change because this is
how the to-pushremote push was done prior to 7fee4089f (Refactor
commands that can set the push-remote before their main task,
2019-03-29).

Re: #3847
Fixes #3872.