magit-branch: Allow it to specify a branch when deleting the current
While working on a topic branch, it's not the master branch I want to
go back to when I delete the current branch, but the topic branch I'm
working on.  This often happens when the current branch is just a
throw away branch for a build test at the middle of the topic branch.

This commit add another option to magit-branch-delete so that we can
specify a branch to checkout when deleting the current branch.

Signed-off-by: Yasushi SHOJI <yasushi.shoji@gmail.com>