magit-skip-worktree-files: Return paths relative to top-level
Return paths relative to the top-level of the repository (which Git
calls "full" name) instead of relative to the `default-directory'.

This fixes bugs in the callers `magit-skip-worktree' and
`magit-no-skip-worktree' when they are called from a sub-directory.
The former used to offer invalid choices and the latter failed to
offer any choices when called from a sub-directory.