magit-generate-changelog: New command
Also suggest it as an option for `git-commit-setup-hook', but
don't use it by default because that needs Emacs 27, doesn't
work well when amending, and would be intrusive when using a
different commit message style.

Add the enhanced paragraph filling of ChangeLog entries into
`git-commit-setup-changelog-support' though, which is useful
when when not using the new command.

Closes #3928.

Modified-by: Jonas Bernoulli <jonas@bernoul.li>
  to merge two commits, move the command to a different file,
  drop some function declarations, and change the commit message.