Bind RET to bug-reference-push-button
... but only if the button is inside a Magit buffer.  In Magit buffers
users expect "RET" to visit the thing at point, so this should happen
for bug-reference buttons too.  However in file-visiting buffers "RET"
is expected to insert a newline character and we shouldn't mess with
that.  By remapping `magit-visit-thing' instead of binding "RET"
explicitly we can limit the effect to Magit buffers.

Closes #3933.