magit-diff-visit-file: Visit the blob for uncommitted removals
Previously this visited to the file in the working tree.  There was
no way to visit the HEAD blob instead.  The HEAD blob is the last
blob that still contains the removed lines and going there is useful
because it allows blaming those lines.

Users who don't like this can set `magit-diff-visit-avoid-head-blob'.

Re #3848.