Indicated if a log section might be abbreviated
For performance reasons log sections by default only show up to 256
commits.  The user can also set another limit N.  If a log section
shows exactly N commits, then that most likely means that there are
more commits and that the limit has kicked in.  It could also mean
that there are exactly N commits, but that is much less likely and
telling the difference would come at a cost, which goes against the
goal of making things faster.  In either case show the number of
commits as "N+" (as in "N or greater") instead of as "N".

Closes #3941.