magit-disable-line-numbers: New option
Some users who turn on `global-disable-line-numbers-mode' (or
`global-nlinum-mode' or `global-linum-mode') expect line numbers
to displayed everywhere except in Magit buffers.  Other users do
not expect Magit buffers to be treated differently.

At least in theory users in the first group should not use the global
mode, but that ship has sailed.  But expecting the global mode to
have a global effect should also be possible.  In fact it should be
the default, but for historic reasons I am keeping the old behavior as
the default for now.  But I am considering changing the default value
before the next release.

Closes #3971.

Some relevant commits:
1: a3eafea1cfaa7b60b8d73b2ee42f297b6542eb4a from 2015
   magit-mode: turn off linum-mode
2: 6e8b7fc1f6493971d6786bd0f13a75417d6a334d from 2016
   Only turn (n)linum-mode off in global-(n)linum-mode
3: 802fba7e0a8856a6d369f49d6610e3ecb0240f00 from 2018
   Turn off display-line-numbers-mode if on