Allow the directory at point as gitignore rule
In the magit status buffer, move the point over an untracked directory, that
is, a directory with only (and at least one) untracked files. Pressing "ip"
to add an ignore rule should probably suggest that directory.  This commit
makes it so.

Previously only untracked files were suggested, and globs of files
with the same extension as untracked files.  This extends the
gitignore suggestions by adding untracked directories.

Modified-by: Kyle Meyer <kyle@kyleam.com>