Explicitly :extend faces for Emacs 27
For some of these faces it isn't strictly necessary to do so because
they inherit from faces that already do, but it seems like a good idea
to do so anyway, if for no other purpose then to serve as some sort of
documentation.

A few other faces are explicitly extended even though they by default
do not even set an attribute that is effected by this.  This is done
so that a user can set such an attribute without subsequently having
to figure out why that doesn't extend to the edge of the window.

See https://debbugs.gnu.org/cgi/bugreport.cgi?bug=37774
and https://github.com/magit/magit/issues/3986.