Adjust copy-region-as-kill calls to fix append behavior
In a few copy commands, we fall back to copy-region-as-kill when the
region is active, calling the function with nil for its BEG and END
and with a non-nil REGION.  According to its doc-string, the behavior
in this case is to "ignore BEG and END, and save the current region
instead".  However, when going down the kill-append code path of
copy-region-as-kill, BEG and END are still used and must be an integer
to avoid a type error.

Update these commands to execute copy-region-as-kill with
call-interactively so that the arguments are set according to the
interactive spec.

Fixes #4001.