magit-ignore-submodules-p: Honor diff.ignoreSubmodules
It used to be possible to ignore submodules only in some repositories
by setting `magit-diff-section-arguments' and `magit-diff-arguments'.
These Lisp variables no longer exist and the user now has to set the
Git variable `diff.ignoreSubmodules' instead.

This function exists because some parts of Magit that don't actually
call "git diff" also want to respect this setting.  To do so properly
we have to consider the variable as well.