magit-copy-revision-abbreviated: New option
magit-copy-buffer-revision, and magit-copy-section-value when
appropriate, pushes the full revision hash to magit-revision-stack and
the kill ring.  If callers would like to use this to insert an
abbreviated hash, their main options are either to
(1) yank the full hash and truncate it manually or (2) configure
magit-pop-revision-stack-format accordingly and call
magit-pop-revision-stack.

Offer an option to put the abbreviated hash on the kill ring and the
magit-revision-stack.