diff: Fix staging of intent-to-add hunks with recent Git versions
As of Git 2.19.0, specifically 0231ae71d3 (diff: turn
--ita-invisible-in-index on by default, 2018-05-26), `git diff`
(without --cached) displays intent-to-add files as new files rather
than empty files.  This breaks staging a region from an intent-to-add
file: when magit-stage derives a patch based on this output and feeds
it to `git apply --cached`, it fails because the intent-to-add file
_is_ already in the index.

Work around this by passing --ita-visible-in-index when we insert
diffs, restoring the pre-2.19.0 default behavior.  Note that we can't
pass this flag unconditionally because it isn't available until Git
2.11.0.  As suggested by Jonas in gh-4026, it's safe to cache the
value to limit the cost of this version check to one additional git
call per session as opposed to an additional git call per
magit--insert-diff call.  The worst case scenario is the unlikely
event that the user upgrades to 2.19.0 or later during a session, in
which case intent-to-add staging (a relatively obscure feature) won't
be available for that session.

Fixes #4026.