diff: Cache intent-to-add kludge per repository
The last commit argued that caching a _single_ value for the version
check per session is ok because the Git version is unlikely cross the
2.19.0 boundary during a session.  And even if this happens, the only
thing that is affected is intent-to-add staging.  That rationale,
however, doesn't consider that having multiple Git versions within a
session would be common for users that access repositories over Tramp.

Not that, in this scenario, breakage isn't necessarily restricted to
intent-to-add staging not working.  Say the user has Git 2.19.0 or
later on their local machine.  This value is cached, and we pass
--ita-visible-in-index when generating the diff.  If a remote machine
that the user visits later has a Git before 2.11.0, which doesn't
support the --ita-visible-in-index option, magit--insert-diff calls
would fail in that remote repository.

Avoid the above issue by caching the version check for each
repository.

Suggested-by: Phil Sainty <phil@catalyst.net.nz>