No longer define magit-section's hook variables as custom options
Instead define them as regular variables and add the default hook
functions using `add-hook'.

We need to do this so that we can distribute `magit-section' as a
standalone package, which cannot use function that are defined in
`magit's on its hooks--they might not be defined since it is now
possible to load `magit-section' without loading `magit'.