magit-wip-commit-worktree: Avoid failing git call on file deletion
When magit-wip-commit-worktree is passed a file that is deleted in the
working tree, its call to `git add' fails if the file is already
deleted in the temporary index.  This scenario is routine when the
composite magit-wip-mode is enabled rather than one of its components,
because the same working tree is often synced with the same index
multiple times (e.g., after save, then before staging, and then again
after staging).

The failing `git add' call is problematic for two reasons:

  * A user seeing the failure message (in the minibuffer, status
    buffer, or process buffer) would reasonably be concerned that
    there's a problem.

  * The failure to add a deleted-already-in-temporary-index file will
    prevent any of the specified files from being added to the
    temporary index and committed to the wtree ref.

    While this is more serious than the first issue, it's pretty
    difficult to access, particularly when using magit-wip-mode, where
    working tree changes will be committed each time a file is saved.

Address both of these issues by using `update-index' instead, which
will happily accept a deleted working tree file that is already
deleted in the index.

Fixes #4040.