Quick fix for log output with "..." as first line
When running "log --graph -G<...>" or "log --graph --follow --
<file...>", git can output "..." separators.  Magit failed to parse
"..." as first line.

This fix now displays as-is any line that doesn't match the expected
regexp.

This is only a quick fix because "..." are still not formatted
properly: the first "..." is not aligned, and subsequent ones are
displayed twice.

Note: I am not sure why git decides to output "..." separators, for
example: "git log --graph -- file" is fine, but add "--follow" and
then you have "..." everywhere, even if the shown histories are nearly
the same.

Fixes #4048.