magit--wip-ref: Fix regression in handling of HEAD
404bd1148 (magit--wip{,-index,-wtree}-ref: New functions, 2018-08-02)
introduced magit--wip-ref to handle the construction of wip refs.  For
a detached head, the wip ref should be ".../{wtree,index}/HEAD".
magit--wip-ref returns the correct wip ref for a detached HEAD when
REF is nil.  However, it incorrectly constructs the ref as
"{wtree,index}/refs/heads/HEAD" when "HEAD" is passed as REF.  As a
result, commits for a detached HEAD are made on
"{wtree,index}/refs/heads/HEAD".

Add a condition to magit--wip-wtree-ref to avoid treating a REF of
"HEAD" as a regular branch.