magit-patch-create: Add options for populating patch text
These two options are convenient for automatically generating
different parts of the patch text.

Hide --cover-from-description because the default mode of "message" is
probably sufficient for most callers.  Hide --notes by default
because, while convenient, it's not clear it'd be used by more than a
small subset of users (of the already small subset using format-patch
often).