magit-patch-create: Add --{inter,range-}diff
Interdiffs and range-diffs are used to help reviewers digest the
changes across iterations of a series.  The --interdiff and
--range-diff options for format-patch are convenient because they save
the sender from having to manually run a command (git-diff or
git-range-diff, respectively) and insert the output into the cover
letter.

Add both options, but hide --interdiff because --range-diff is more
versatile and provides more information.