magit-autorevert: Don't ask if .git/ files are tracked
When magit-auto-revert-tracked-only is non-nil, visiting a file in
.git/ (e.g., COMMIT_EDITMSG) leads to
magit-turn-on-auto-revert-mode-if-desired asking whether the file is
tracked by running `ls-files --error-unmatch' on the file from within
.git/.

While that ls-files query is pointless (we already know files under
.git/ are untracked), it should be safe to assume that ls-files will
always give a non-zero exit in this case.  Due to a bug in the most
recent Git release candidate, the above ls-files call leads to a
segfault [0], but this is fixed with 5c20398699 (prefix_path: show
gitdir if worktree unavailable, 2020-03-02).

... justify why this is worth it ...

[0]: https://lore.kernel.org/git/20200228195805.GA190372@google.com/
     https://lore.kernel.org/git/87h7yr8omj.fsf@kyleam.com/