magit-reshelve-since: Avoid git-filter-branch warning
As of Git v2.24.0, specifically 9df53c5de6 (Recommend git-filter-repo
instead of git-filter-branch, 2019-09-04), Git recommends against
git-filter-branch (and suggests the out-of-tree git-filter-repo as an
alternative).  Each invocation of git-filter-branch has a 10 second
delay where it shows a warning.

Based on reading the (long) list of issues in the git-filter-branch
manpage, I don't spot any that I think apply to the restricted fashion
in which we're using filter-branch in magit-reshelve-since.  Set
FILTER_BRANCH_SQUELCH_WARNING to dismiss the warning and avoid the
delay.

If it turns out that we do want to move away from using
git-filter-branch, refs/pull/4090/head provides an implementation that
uses git-fast-{export,import}.

Re: #4090