magit-reshelve-since: Conditionally use fast-{export,import}
As of Git v2.24.0, Git warns against using git-filter-branch:

  WARNING: git-filter-branch has a glut of gotchas generating mangled history
           rewrites.  Hit Ctrl-C before proceeding to abort, then use an
           alternative filtering tool such as 'git filter-repo'
           (https://github.com/newren/git-filter-repo/) instead.  See the
           filter-branch manual page for more details; to squelch this warning,
           set FILTER_BRANCH_SQUELCH_WARNING=1.

Based on reading the (long) list of issues in the git-filter-branch
manpage, I don't spot any that I think apply to the restricted fashion
in which we're using filter-branch in magit-reshelve-since.  So, if it
were difficult to use another approach or required installing an
external tool, it might make sense to stick with filter-branch.

However, we can pretty easily and cleanly avoid filter-branch by
dumping git-fast-export output to a temporary buffer, editing the
timestamps, and then bringing the changes back in with
git-fast-import.  Unfortunately, we still have to keep the
filter-branch code path around to fall back to because the
git-fast-export call needs --reference-excluded-parents, which didn't
exist until Git v.2.21.0.

Note that there is a subtle change in behavior in how mismatches
between the input time zone and the local time zone are handled.  Say
the local environment is in time zone A and a time with time zone B is
entered into the prompt.  In both implementations, the unix epoch
timestamp is calculated considering the entered time zone, B.  The
difference comes in whether time zone B is carried through to the new
commits.  In the filter-branch-based implementation, commits are made
with the timestamp and no time zone specified, so Git uses A, the
local time zone.  The fast-export approach, on the other hand, uses
time zone B in the new commits.  So, in this scenario, both approaches
create commits with same unix timestamp, but the contextual timestamp
that Git records is different.