magit-{git,shell}-command-topdir: Complete from top-level
The `magit-shell-command-topdir` and `magit-gitcommand-topdir`
completion is now done in the context of the repository top level,
which allows completing scripts and other files residing there.
Previously the completion was done from the context of the current
`default-directory` which wasn't necessarily the repository top level
making the completion of local files nonsensical.

By using `default-directory` in the let form, things should be
consistent now.

Modified-by: Kyle Meyer <kyle@kyleam.com>
  Relocated the abbreviate-file-name call.