magit-abbrev-length: Handle explicit core.abbrev=auto
magit-abbrev-length considers core.abbrev to be "auto" when
core.abbrev is unset (Git's default as of 2.11) but does not account
for an explicit value of "auto".  An explicit "auto" is fed to
string-to-number, resulting in magit-abbrev-length incorrectly
returning 0.  The main visible effect of this is misalignment in log
buffers.

Check whether core.abbrev is set to "auto".

Re: #4123