magit-abbrev-length: Try harder for find sample commits
magit-abbrev-length looks at how HEAD and HEAD~ are abbreviated to
guess Git's automatic abbreviation length, falling back to 7 if those
aren't defined.  When HEAD and HEAD~ do not exist, a repository is
usually just starting out, so 7 (what core.abbrev=auto starts at) is a
good answer.

This isn't the correct answer, though, in the uncommon case of being
on an unborn branch in a repository where Git would abbreviate to a
value greater than 7.  Likewise, in the same scenario except for _one_
commit on the current branch, falling back to 7 for HEAD~ gives an
abbreviation length that is too small.

Get a better answer for these cases by (1) taking two sample commits
from the output of 'rev-list --all' when HEAD and HEAD~ both don't
exist and (2) using HEAD's abbreviation length rather than 7 if HEAD~
doesn't exist.

Note that the rev-parse calls on HEAD and HEAD~ could be dropped,
relying on only the rev-list call, but doing that would impose a
performance cost on all magit-abbrev-length calls in order to handle a
uncommon scenario.

Closes #4123.