magit-status-jump: Use "Recent commits" description when appropriate
If there are no commits that are unpushed to the upstream, then
`magit-jump-to-unpushed-to-upstream' instead jumps to the "Recent
commits", which actually are another incarnation of the "Unmerged
into upstream" section.

The problem is that users don't know that, so in that case we have
to adjust the shown description accordingly to let them know, or
rather to save them from having to know.

Closes #4130.