magit-insert-log: Use form feed rather than null byte as delimiter
db97623a (magit-insert-{log,bisect-rest}: handle refs with
parentheses, 2017-04-15) introduced the first null byte in the
--format value, and since then the format string has continued to gain
null bytes.  Unfortunately, this interferes with the graph prefix that
Git adds to each line of the text:

    $ git log --graph 9423edc0~2..9423edc0 --format="%h%x00%ad%n%s%n"
    * 9423edc0^@Fri May 22 17:42:34 2020 +0200
    magit-commit-absorb: New command
    |
    * 1f8ee45d^@Fri May 22 17:38:12 2020 +0200
    Rename magit-commit-absorb to magit-commit-autofixup

If the %x00 is replaced with another character, a "| " is appended in
front of magit-commit-absorb as well.

The visible consequence of this is a regression in the display of the
++header output when combined with --graph, where the
magit-log-revision-headers-format lines lack a graph prefix.

Restore the desired output by using a form feed instead of the null
byte.

Closes #4129.